requirements.txt
spacy==3.7.4
python-dotenv
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
deepgram-sdk==2.12.0
requests==2.31.0
python-multipart==0.0.9
fastapi==0.109.2
uvicorn==0.27.1
websockets
httpx
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz